Beep only for Windows

+++ chip8/src/driver.rs
@@ -16,7 +16,7 @@ pub const KEY_E: u8 = 0xE;
 pub const KEY_F: u8 = 0xF;
 
 pub trait Driver {
-    fn sound_do_beep(&mut self, frequency: u16, duration: f64);
+    fn sound_do_beep(&mut self, frequency: u32, duration: u32);
     fn video_fill_buffer(&mut self, display: &Vec<Vec<usize>>);
     fn input_is_key_down(&mut self, key: u8) -> bool;
     fn input_is_key_up(&mut self, key: u8) -> bool;
